Use double-precision coordinates
All coordinates are now double-precision. The test data file has been
updated to list coordinates as double-precision floating point numbers.

+++ cell.cpp
@@ -17,7 +17,7 @@ bool Cell::cellHasVertex( const std::string& vertexName ) const
     return false;
 }
 
-bool Cell::cellHasVertex( const int x, const int y ) const
+bool Cell::cellHasVertex( const double x, const double y ) const
 {
     for ( const auto& vertex : vertices ) {
         if ( vertex->hasSameCoordinates( x, y ) ) return true;
